backend/api/runtime/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile backend/api/runtime/requirements.in
#
aws-lambda-powertools==1.29.2
    # via -r backend/api/runtime/requirements.in
aws-xray-sdk==2.10.0
    # via aws-lambda-powertools
boto3==1.24.80
    # via
    #   -r backend/api/runtime/requirements.in
    #   aws-lambda-powertools
botocore==1.27.80
    # via
    #   aws-xray-sdk
    #   boto3
    #   s3transfer
fastjsonschema==2.16.2
    # via aws-lambda-powertools
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
python-dateutil==2.8.2
    # via botocore
s3transfer==0.6.0
    # via boto3
six==1.16.0
    # via python-dateutil
urllib3==1.26.12
    # via botocore
wrapt==1.14.1
    # via aws-xray-sdk
python-dotenv~=0.21.0
bs4~=0.0.1
beautifulsoup4~=4.11.1
